+++ package.json
@@ -12,7 +12,7 @@
 		"lint-fix": "biome lint --write --no-errors-on-unmatched --files-ignore-unknown=true .",
 		"format": "biome format --no-errors-on-unmatched --files-ignore-unknown=true --write .",
 		"type-check": "tsc --noEmit -p tsconfig.json",
-		"test": "vitest",
+		"test": "vitest --no-watch",
 		"prepare": "if [ -z \"$CI\" ]; then husky; fi",
 		"generate-images": "node --loader ts-node/esm scripts/svg-to-png.ts"
 	},
@@ -33,6 +33,7 @@
 		"html-minifier": "^4.0.0",
 		"jsdom": "^24.1.3",
 		"libphonenumber-js": "^1.11.19",
+		"pg": "^8.13.3",
 		"sharp": "^0.33.5",
 		"tailwindcss": "^3.4.16",
 		"vue": "^3.5.13"
