+++ src/components/SignUpForm.astro
@@ -126,6 +126,9 @@ declare global {
 	const animationManager = createAnimationManager();
 	createAutoAdvanceManager(); // We don't call this anywhere, but it needs to be initialized
 	
+	// Get the default button text
+	const defaultText = "Sign Up for Notifications";
+
 	// Expose Turnstile callback to window
 	window.onTurnstileCallback = (token: string): void => {
 		const form = document.getElementById("signup-form");
@@ -202,9 +205,17 @@ declare global {
 			// Handle error response
 			animationManager.cleanup();
 			
+			// Store the error message
+			const errorMessage = button.textContent?.trim();
+			
 			// Reset button after delay
 			setTimeout(() => {
 				buttonStateManager.reset();
+				// Restore the error message
+				if (errorMessage && errorMessage !== defaultText) {
+					button.innerHTML = errorMessage;
+					button.setAttribute('data-error', 'true');
+				}
 				animationManager.initHoverAnimations();
 			}, 5000);
 		}
