+++ src/components/contact-info/UserContactInfo.astro
@@ -2,8 +2,8 @@
 import { neon } from "@neondatabase/serverless";
 import PhoneInput from "../contact-info/PhoneInput.vue";
 import CityInput from "../contact-info/CityInput.vue";
-import type { City, CityOption } from "../../../shared/types/form.schema";
-import { CONTACT_SCHEMA } from "../../../shared/types/form.schema";
+import type { City, CityOption } from "../../../shared/form.schema";
+import { CONTACT_SCHEMA } from "../../../shared/form.schema";
 
 const sql = neon(import.meta.env.DATABASE_URL);
 
@@ -47,14 +47,14 @@ const cityOptions: CityOption[] = cities.map((city) => ({
 	<div class="grid grid-cols-1 gap-4">
 		<div class="relative">
 			<label for="name" class="block text-sm font-medium text-slate-700 mb-1">
-				{CONTACT_SCHEMA.name.formLabel}
+				{CONTACT_SCHEMA.preferred_name.form_label}
 			</label>
 			<input
 				type="text"
 				id="name"
 				name="name"
-				placeholder={CONTACT_SCHEMA.name.placeholder}
-				required={CONTACT_SCHEMA.name.required}
+				placeholder={CONTACT_SCHEMA.preferred_name.placeholder}
+				required={CONTACT_SCHEMA.preferred_name.required}
 				class="w-full px-4 py-2 border border-slate-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:border-indigo-500"
 			/>
 		</div>
