+++ src/components/notifications/Notifications.astro
@@ -2,6 +2,17 @@
 import NotificationCategory from "./NotificationCategory.astro";
 import NotificationItem from "./NotificationItem.astro";
 import UserPreferences from "../user-preferences/UserPreferences.astro";
+import { NOTIFICATION_SCHEMA } from "../../../shared/form.schema";
+import {
+	groupNotificationsByCategory,
+	getNotificationCategories,
+	CATEGORY_DESCRIPTIONS,
+	type NotificationWithMetadata,
+} from "../../../shared/notifications";
+
+const notificationsByCategory =
+	groupNotificationsByCategory(NOTIFICATION_SCHEMA);
+const categories = getNotificationCategories(NOTIFICATION_SCHEMA);
 ---
 
 <section>
@@ -14,7 +25,7 @@ import UserPreferences from "../user-preferences/UserPreferences.astro";
     </h2>
     
     <!-- Badge showing number of notification types -->
-    <span class="bg-indigo-100 text-indigo-800 text-xs font-medium px-2.5 py-1 rounded-full">5 types available</span>
+    <span class="bg-indigo-100 text-indigo-800 text-xs font-medium px-2.5 py-1 rounded-full">{Object.keys(NOTIFICATION_SCHEMA).length} types available</span>
   </div>
   
   <div class="bg-indigo-50/50 p-4 rounded-lg border border-indigo-100 mb-6">
@@ -23,67 +34,26 @@ import UserPreferences from "../user-preferences/UserPreferences.astro";
   
   <!-- Two-column layout for notification categories -->
   <div class="grid grid-cols-1 lg:grid-cols-2 gap-8">
-    <!-- Nature & Sky Events Category -->
-    <NotificationCategory 
-      title="Nature & Sky Events" 
-      description="Experience the awe of the universe and never miss nature's most spectacular moments"
-    >
-      <div class="grid grid-cols-1 gap-3">
-        <NotificationItem 
-          id="nasa"
-          value="nasa"
-          label="NASA Image of the Day"
-          description="Wake up to the wonders of space. Each day brings a breathtaking new view of our universe, from distant galaxies to our own cosmic backyard."
-          badgeType="daily"
-          imageUrl="/assets/notifications/astrological.webp"
-        />
-        
-        <NotificationItem 
-          id="sunset-alert"
-          value="sunset-alert"
-          label="Golden Hour Alerts"
-          description="Never miss a magical sunset again. Get notified 30 minutes before the perfect golden hour for stunning photos and peaceful moments."
-          badgeType="instant"
-          imageUrl="/assets/notifications/sunset.jpg"
-        />
-        
-        <NotificationItem 
-          id="full-moon"
-          value="full-moon"
-          label="Celestial Events & Night Sky"
-          description="Never miss the magic above. Get alerts for full moons, meteor showers, eclipses, planetary alignments, and rare astronomical events perfect for stargazing and photography."
-          badgeType="scheduled"
-          imageUrl="/assets/notifications/celestial.webp"
-        />
-      </div>
-    </NotificationCategory>
-    
-    <!-- Daily Life Category -->
-    <NotificationCategory 
-      title="Daily Life" 
-      description="Smart notifications that make your day better"
-      isLastCategory={true}
-    >
-      <div class="grid grid-cols-1 gap-3">
-        <NotificationItem 
-          id="weather-outfit"
-          value="weather-outfit"
-          label="Smart Weather & Style"
-          description="Start your day with confidence. Get personalized weather updates and outfit suggestions based on the forecast."
-          badgeType="daily"
-          imageUrl="/assets/notifications/weather-example.jpg"
-        />
-        
-        <NotificationItem 
-          id="daily-recipe"
-          value="daily-recipe"
-          label="Daily Recipe Inspiration"
-          description="Break free from cooking ruts. Receive chef-curated recipes with easy-to-find ingredients and step-by-step instructions."
-          badgeType="daily"
-          imageUrl="/assets/notifications/recipe-example.jpg"
-        />
-      </div>
-    </NotificationCategory>
+    {categories.map((category: string, index: number) => (
+      <NotificationCategory 
+        title={CATEGORY_DESCRIPTIONS[category].title}
+        description={CATEGORY_DESCRIPTIONS[category].description}
+        isLastCategory={index === categories.length - 1}
+      >
+        <div class="grid grid-cols-1 gap-3">
+          {notificationsByCategory[category].map((notification: NotificationWithMetadata) => (
+            <NotificationItem 
+              id={notification.id}
+              value={notification.id}
+              label={notification.formLabel}
+              description={notification.metadata.description}
+              badgeType={notification.metadata.badgeType}
+              imageUrl={notification.metadata.imageUrl}
+            />
+          ))}
+        </div>
+      </NotificationCategory>
+    ))}
   </div>
 
   <!-- Preferences section -->
