+++ src/components/user-preferences/UserPreferences.astro
@@ -1,5 +1,5 @@
 ---
-import { PREFERENCES_SCHEMA } from "../../../shared/types/form.schema";
+import { PREFERENCES_SCHEMA } from "../../../shared/form.schema";
 ---
 
 <section class="user-preferences bg-white/90 backdrop-blur-sm rounded-xl border border-indigo-100 overflow-hidden" x-data="{ open: false }">
@@ -46,17 +46,17 @@ import { PREFERENCES_SCHEMA } from "../../../shared/types/form.schema";
             <!-- Language preference -->
             <div>
                 <label for="preferred-language" class="block text-sm text-slate-600 mb-1">
-                    {PREFERENCES_SCHEMA.preferredLanguage.formLabel}
+                    {PREFERENCES_SCHEMA.preferred_language.form_label}
                 </label>
                 <div class="relative">
                     <select 
                         id="preferred-language"
                         name="preferred-language"
-                        aria-label={PREFERENCES_SCHEMA.preferredLanguage.formLabel}
+                        aria-label={PREFERENCES_SCHEMA.preferred_language.form_label}
                         aria-describedby="preferred-language-description"
                         class="w-full px-3 py-2 bg-white border border-slate-200 rounded-md focus:outline-none focus:ring-1 focus:ring-indigo-500 text-sm text-slate-800 appearance-none"
                     >
-                        {Object.values(PREFERENCES_SCHEMA.preferredLanguage.options).map(option => (
+                        {Object.values(PREFERENCES_SCHEMA.preferred_language.options).map(option => (
                             <option value={option.code}>
                                 {option.name}
                             </option>
@@ -73,17 +73,17 @@ import { PREFERENCES_SCHEMA } from "../../../shared/types/form.schema";
             <!-- Unit preference -->
             <div>
                 <label for="unit-preference" class="block text-sm text-slate-600 mb-1">
-                    {PREFERENCES_SCHEMA.unitPreference.formLabel}
+                    {PREFERENCES_SCHEMA.unit_preference.form_label}
                 </label>
                 <div class="relative">
                     <select 
                         id="unit-preference"
                         name="unit-preference"
-                        aria-label={PREFERENCES_SCHEMA.unitPreference.formLabel}
+                        aria-label={PREFERENCES_SCHEMA.unit_preference.form_label}
                         aria-describedby="unit-preference-description"
                         class="w-full px-3 py-2 bg-white border border-slate-200 rounded-md focus:outline-none focus:ring-1 focus:ring-indigo-500 text-sm text-slate-800 appearance-none"
                     >
-                        {Object.values(PREFERENCES_SCHEMA.unitPreference.options).map(option => (
+                        {Object.values(PREFERENCES_SCHEMA.unit_preference.options).map(option => (
                             <option value={option.code}>
                                 {option.name}
                             </option>
@@ -100,17 +100,17 @@ import { PREFERENCES_SCHEMA } from "../../../shared/types/form.schema";
             <!-- Time Format -->
             <div>
                 <label for="time-format" class="block text-sm text-slate-600 mb-1">
-                    {PREFERENCES_SCHEMA.timeFormat.formLabel}
+                    {PREFERENCES_SCHEMA.time_format.form_label}
                 </label>
                 <div class="relative">
                     <select 
                         id="time-format"
                         name="time-format"
-                        aria-label={PREFERENCES_SCHEMA.timeFormat.formLabel}
+                        aria-label={PREFERENCES_SCHEMA.time_format.form_label}
                         aria-describedby="time-format-description"
                         class="w-full px-3 py-2 bg-white border border-slate-200 rounded-md focus:outline-none focus:ring-1 focus:ring-indigo-500 text-sm text-slate-800 appearance-none"
                     >
-                        {Object.values(PREFERENCES_SCHEMA.timeFormat.options).map(option => (
+                        {Object.values(PREFERENCES_SCHEMA.time_format.options).map(option => (
                             <option value={option.code}>
                                 {option.name}
                             </option>
@@ -127,17 +127,17 @@ import { PREFERENCES_SCHEMA } from "../../../shared/types/form.schema";
             <!-- Daily Notification Time -->
             <div>
                 <label for="daily-notification-time" class="block text-sm text-slate-600 mb-1">
-                    {PREFERENCES_SCHEMA.dailyNotificationTime.formLabel}
+                    {PREFERENCES_SCHEMA.daily_notification_time.form_label}
                 </label>
                 <div class="relative">
                     <select 
                         id="daily-notification-time"
                         name="daily-notification-time"
-                        aria-label={PREFERENCES_SCHEMA.dailyNotificationTime.formLabel}
+                        aria-label={PREFERENCES_SCHEMA.daily_notification_time.form_label}
                         aria-describedby="daily-notification-time-description"
                         class="w-full px-3 py-2 bg-white border border-slate-200 rounded-md focus:outline-none focus:ring-1 focus:ring-indigo-500 text-sm text-slate-800 appearance-none"
                     >
-                        {Object.values(PREFERENCES_SCHEMA.dailyNotificationTime.options).map(option => (
+                        {Object.values(PREFERENCES_SCHEMA.daily_notification_time.options).map(option => (
                             <option value={option.code}>
                                 {option.name}
                             </option>
@@ -155,7 +155,7 @@ import { PREFERENCES_SCHEMA } from "../../../shared/types/form.schema";
 </section>
 
 <script>
-    import { IMPERIAL_COUNTRIES, LANGUAGE_OPTIONS } from "../../../shared/types/form.schema";
+    import { IMPERIAL_COUNTRIES, LANGUAGE_OPTIONS } from "../../../shared/form.schema";
     import Alpine from 'alpinejs';
     
     // Declare Alpine on window
